feat: holy molly build rug
