+++ package.json
@@ -86,9 +86,6 @@
     "@metamask/detect-provider": "^1.2.0",
     "@react-spring/web": "^9.5.2",
     "@reduxjs/toolkit": "^1.9.3",
-    "@safe-global/api-kit": "^2.4.3",
-    "@safe-global/protocol-kit": "^4.0.3",
-    "@safe-global/safe-core-sdk-types": "^5.0.3",
     "@sentry-internal/browser-utils": "^8.0.0-rc.2",
     "@sentry/react": "^7.112.2",
     "@shapeshiftoss/caip": "workspace:^",
